feat(SwiftDeclarationRendering): reader-specialize FieldLayoutRenderer
Split FieldLayoutRenderer into a thin generic facade dispatching (via
`self as? FieldLayoutRenderer<MachOImage>/<MachOFile>`) to two reader-specialized
implementations, so the same field-comment rendering works both in-process and
offline:

- FieldLayoutRenderer+MachOImage.swift: the existing runtime path (in-process
  metadata: createInProcess / value-witness tables / getTypeByMangledNameInContext),
  moved verbatim with the entry points renamed. The former +Enum file folds in here.
- FieldLayoutRenderer+MachOFile.swift: a new static path computing field offsets,
  end offsets, type layouts, the expanded nested-offset tree, and enum layouts
  from the binary via SwiftLayout — never loading the process.

Injection seam: StaticFieldLayoutProvider (a reader-agnostic protocol carried in
the non-generic DeclarationRenderConfiguration) wraps a StaticLayoutCalculator,
built once per session and injected, so the (relatively expensive) dependency
closure is not rebuilt per type. Resolution defaults to the transitive
dependency closure over the system dyld shared cache; `.singleImage` restricts it.

Without a provider the offline path emits nothing, exactly as before (offline
metadata was unavailable, so the runtime path produced no comments either) — no
regression. SwiftDeclarationRendering now depends on SwiftLayout (acyclic).

Note: typeLayoutTransformer applies only to the runtime path (it is typed on the
runtime TypeLayout, which cannot be synthesized from the static TypeLayoutInfo
outside MachOSwiftSection); the static path always emits the default format.
